cub3d

+++ cub3d/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: rvena <rvena@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/18 17:37:47 by rvena             #+#    #+#             */
-/*   Updated: 2021/04/08 17:17:15 by rvena            ###   ########.fr       */
+/*   Updated: 2021/04/08 18:40:22 by rvena            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -105,8 +105,8 @@ typedef struct set_data
 	char	*WE;
 	char	*EA;
 	char	*S;
-	int		floor;
-	int		ceilling;
+	long	floor;
+	long	ceilling;
 	int		mapW;
 	int		mapH;
 }				settings;
